fix - Intercom initializer
