+++ src/hooks/useWalletState.ts
@@ -2,9 +2,6 @@ import React, { useEffect } from 'react';
 import { getENSAddress, smallWalletAddress } from '@src/model/wallet';
 import { globalWalletState } from '@stores/WalletStore';
 import { useState } from '@hookstate/core';
-import crypto from 'crypto';
-import { INTERCOM_KEY } from '@env';
-import Intercom from '@intercom/intercom-react-native';
 
 const useWalletState = () => {
 	const state = useState(globalWalletState());
@@ -21,20 +18,9 @@ const useWalletState = () => {
 	}, [address]);
 
 	const accountName = () => {
-		if (ensName) {
-			return ensName;
-		}
+		if (ensName) return ensName;
 		return smallWalletAddress(address);
 	};
-
-	useEffect(() => {
-		const intercomKey = INTERCOM_KEY || process.env.INTERCOM_KEY;
-		const hmac = crypto.createHmac('sha256', intercomKey!);
-		hmac.update(accountName());
-		const sign = hmac.digest('hex');
-		Intercom.setUserHash(sign);
-		Intercom.registerIdentifiedUser({ userId: accountName() });
-	}, [address]);
 
 	return {
 		state,
