+++ src/routes/index.tsx
@@ -1,6 +1,9 @@
 import React, { useEffect } from 'react';
 import { useState } from '@hookstate/core';
 import RNUxcam from 'react-native-ux-cam';
+import crypto from 'crypto';
+import { INTERCOM_KEY } from '@env';
+import Intercom from '@intercom/intercom-react-native';
 import { TransactionsProvider } from '@contexts';
 import { useWalletState } from '@hooks';
 import { NavigationContainer } from '@react-navigation/native';
@@ -23,6 +26,8 @@ const Routes: React.FC = () => {
 	RNUxcam.setUserIdentity(accountName);
 
 	const walletState = useState(globalWalletState());
+	const { address } = walletState.value;
+
 	const initialScreen = walletState.value.walletId ? 'HomeScreen' : 'WelcomeScreen';
 
 	const urlRedirect = (event: any) => {
@@ -35,11 +40,18 @@ const Routes: React.FC = () => {
 
 	useEffect(() => {
 		Linking.getInitialURL().then((url) => url && urlRedirect({ url }));
-
 		Linking.addEventListener('url', urlRedirect);
-
 		return () => Linking.removeEventListener('url', urlRedirect);
 	}, []);
+
+	useEffect(() => {
+		const intercomKey = INTERCOM_KEY || process.env.INTERCOM_KEY;
+		const hmac = crypto.createHmac('sha256', intercomKey!);
+		hmac.update(accountName);
+		const sign = hmac.digest('hex');
+		Intercom.setUserHash(sign);
+		Intercom.registerIdentifiedUser({ userId: accountName });
+	}, [address]);
 
 	const linking = {
 		prefixes: [prefix],
